Move GIF commands to actions folder
